Split model tests into separate test files
