+++ api/tests/models/test_event.py
@@ -1,0 +1,10 @@
+from django.test import TestCase
+from api.models import Event
+
+
+class EventModelTest(TestCase):
+  def setUp(self):
+    self.event1 = Event.objects.create(name='event1', games='2016 Summer', sport='Badminton')
+
+  def test_string_representation(self):
+    self.assertEqual(str(self.event1), self.event1.name)
